Separate styles from the main `IOButton` component logic
