+++ src/components/buttons/IOButton/__test__/IOButton.test.tsx
@@ -2,9 +2,7 @@ import React from "react";
 import { Alert } from "react-native";
 import * as TestRenderer from "react-test-renderer";
 import IOButton from "../IOButton";
-import IconButton from "../IconButton";
-import IconButtonContained from "../IconButtonContained";
-import IconButtonSolid from "../IconButtonSolid";
+
 const onButtonPress = () => {
   Alert.alert("Alert", "Action triggered");
 };
@@ -44,38 +42,5 @@ describe("Test Buttons Components", () => {
       ></IOButton>
     ).toJSON();
     expect(buttonOutline).toMatchSnapshot();
-  });
-
-  it("IconButtonSolid Snapshot", () => {
-    const iconButtonSolid = TestRenderer.create(
-      <IconButtonSolid
-        onPress={onButtonPress}
-        icon={"spid"}
-        accessibilityLabel={"accessibilityLabel"}
-      ></IconButtonSolid>
-    ).toJSON();
-    expect(iconButtonSolid).toMatchSnapshot();
-  });
-
-  it("IconButton Snapshot", () => {
-    const iconButton = TestRenderer.create(
-      <IconButton
-        onPress={onButtonPress}
-        icon={"spid"}
-        accessibilityLabel={"accessibilityLabel"}
-      ></IconButton>
-    ).toJSON();
-    expect(iconButton).toMatchSnapshot();
-  });
-
-  it("IconButtonContained Snapshot", () => {
-    const iconButtonContained = TestRenderer.create(
-      <IconButtonContained
-        onPress={onButtonPress}
-        icon={"spid"}
-        accessibilityLabel={"accessibilityLabel"}
-      ></IconButtonContained>
-    ).toJSON();
-    expect(iconButtonContained).toMatchSnapshot();
   });
 });
