+++ src/components/buttons/IOButton/index.tsx
@@ -1,0 +1,1 @@
+export * from "./IOButton";
